Move request package outside of model.
