+++ src/test/java/com/commercetools/project/sync/service/impl/ReferencesServiceImplTest.java
@@ -14,7 +14,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.commercetools.project.sync.model.request.ResourceIdsGraphQlRequest;
+import com.commercetools.project.sync.request.ResourceIdsGraphQlRequest;
 import com.commercetools.project.sync.service.ReferencesService;
 import com.commercetools.sync.commons.models.ResourceKeyId;
 import com.commercetools.sync.commons.models.ResourceKeyIdGraphQlResult;
